+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Reverse complement API has been refactored into plain functions.
+- Reverse complement now supports the whole IUPAC alphabet.
+
+## [0.6.0] - 2016-05-09
+- Type aliases for various text representations.
+- Pattern matching algorithms take both iterators and slices where possible.
+- logprobs::cumsum has been refactored to return an iterator.
+- support for subtraction of logprobs.
 
 ## [0.5.0] - 2016-02-24
 ### Added